Moving tests into the src directory
