+++ .php_cs
@@ -13,7 +13,6 @@ EOF;
 
 $finder = PhpCsFixer\Finder::create()
     ->in(__DIR__.'/src')
-    ->in(__DIR__.'/tests')
     ->in(__DIR__.'/bin')
 ;
