+++ phpstan.neon
@@ -1,5 +1,7 @@
 includes:
   - vendor/phpstan/phpstan-strict-rules/rules.neon
+  - vendor/phpstan/phpstan-phpunit/extension.neon
+  - vendor/phpstan/phpstan-phpunit/rules.neon
 parameters:
   checkMissingIterableValueType: false
   checkGenericClassInNonGenericObjectType: false
