--- phpstan.tests.neon
@@ -1,7 +1,0 @@
-includes:
-  - vendor/phpstan/phpstan-strict-rules/rules.neon
-  - vendor/phpstan/phpstan-phpunit/extension.neon
-  - vendor/phpstan/phpstan-phpunit/rules.neon
-parameters:
-  checkMissingIterableValueType: false
-  ignoreErrors:
