+++ src/RulesTest.php
@@ -13,16 +13,8 @@
 
 declare(strict_types=1);
 
-namespace Pdp\Tests;
+namespace Pdp;
 
-use Pdp\Domain;
-use Pdp\InvalidDomainName;
-use Pdp\InvalidHost;
-use Pdp\PublicSuffix;
-use Pdp\ResolvedDomain;
-use Pdp\Rules;
-use Pdp\UnableToLoadPublicSuffixList;
-use Pdp\UnableToResolveDomain;
 use PHPUnit\Framework\TestCase;
 use TypeError;
 use function array_fill;
