+++ src/Storage/PublicSuffixListRemoteStorageTest.php
@@ -13,14 +13,13 @@
 
 declare(strict_types=1);
 
-namespace Pdp\Tests\Storage;
+namespace Pdp\Storage;
 
 use GuzzleHttp\Exception\ConnectException;
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
-use Pdp\Storage\RootZoneDatabaseRemoteStorage;
-use Pdp\TopLevelDomains;
-use Pdp\UnableToLoadRootZoneDatabase;
+use Pdp\Rules;
+use Pdp\UnableToLoadPublicSuffixList;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Client\ClientInterface;
 use Psr\Http\Message\RequestFactoryInterface;
@@ -30,18 +29,17 @@ use function dirname;
 use function file_get_contents;
 
 /**
- * @coversDefaultClass \Pdp\Storage\RootZoneDatabaseRemoteStorage
+ * @coversDefaultClass \Pdp\Storage\PublicSuffixListRemoteStorage
  */
-final class RootZoneDatabaseRemoteStorageTest extends TestCase
+final class PublicSuffixListRemoteStorageTest extends TestCase
 {
-    public function testIsCanReturnARootZoneDatabaseInstance(): void
+    public function testIsCanReturnAPublicSuffixListInstance(): void
     {
         $client = new class() implements ClientInterface {
             public function sendRequest(RequestInterface $request): ResponseInterface
             {
                 /** @var string $body */
-                $body = file_get_contents(dirname(__DIR__, 2).'/test_data/tlds-alpha-by-domain.txt');
-
+                $body = file_get_contents(dirname(__DIR__, 2).'/test_data/public_suffix_list.dat');
                 return new Response(200, [], $body);
             }
         };
@@ -53,10 +51,10 @@ final class RootZoneDatabaseRemoteStorageTest extends TestCase
             }
         };
 
-        $storage = new RootZoneDatabaseRemoteStorage($client, $requestFactory);
-        $rzd = $storage->getByUri('http://www.example.com');
+        $storage = new PublicSuffixListRemoteStorage($client, $requestFactory);
+        $psl = $storage->getByUri('http://www.example.com');
 
-        self::assertInstanceOf(TopLevelDomains::class, $rzd);
+        self::assertInstanceOf(Rules::class, $psl);
     }
 
     public function testItWillThrowIfTheClientCanNotConnectToTheRemoteURI(): void
@@ -75,10 +73,10 @@ final class RootZoneDatabaseRemoteStorageTest extends TestCase
             }
         };
 
-        self::expectException(UnableToLoadRootZoneDatabase::class);
-        self::expectExceptionMessage('Could not access the Root Zone Database URI: `http://www.example.com`.');
+        self::expectException(UnableToLoadPublicSuffixList::class);
+        self::expectExceptionMessage('Could not access the Public Suffix List URI: `http://www.example.com`.');
 
-        $storage = new RootZoneDatabaseRemoteStorage($client, $requestFactory);
+        $storage = new PublicSuffixListRemoteStorage($client, $requestFactory);
         $storage->getByUri('http://www.example.com');
     }
 
@@ -98,11 +96,11 @@ final class RootZoneDatabaseRemoteStorageTest extends TestCase
             }
         };
 
-        self::expectException(UnableToLoadRootZoneDatabase::class);
-        self::expectExceptionMessage('Invalid response from Root Zone Database URI: `http://www.example.com`.');
+        self::expectException(UnableToLoadPublicSuffixList::class);
+        self::expectExceptionMessage('Invalid response from Public Suffix List URI: `http://www.example.com`.');
         self::expectExceptionCode(404);
 
-        $storage = new RootZoneDatabaseRemoteStorage($client, $requestFactory);
+        $storage = new PublicSuffixListRemoteStorage($client, $requestFactory);
         $storage->getByUri('http://www.example.com');
     }
 }
