+++ src/Storage/RootZoneDatabaseRemoteStorageTest.php
@@ -13,14 +13,13 @@
 
 declare(strict_types=1);
 
-namespace Pdp\Tests\Storage;
+namespace Pdp\Storage;
 
 use GuzzleHttp\Exception\ConnectException;
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
-use Pdp\Rules;
-use Pdp\Storage\PublicSuffixListRemoteStorage;
-use Pdp\UnableToLoadPublicSuffixList;
+use Pdp\TopLevelDomains;
+use Pdp\UnableToLoadRootZoneDatabase;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Client\ClientInterface;
 use Psr\Http\Message\RequestFactoryInterface;
@@ -30,17 +29,18 @@ use function dirname;
 use function file_get_contents;
 
 /**
- * @coversDefaultClass \Pdp\Storage\PublicSuffixListRemoteStorage
+ * @coversDefaultClass \Pdp\Storage\RootZoneDatabaseRemoteStorage
  */
-final class PublicSuffixListRemoteStorageTest extends TestCase
+final class RootZoneDatabaseRemoteStorageTest extends TestCase
 {
-    public function testIsCanReturnAPublicSuffixListInstance(): void
+    public function testIsCanReturnARootZoneDatabaseInstance(): void
     {
         $client = new class() implements ClientInterface {
             public function sendRequest(RequestInterface $request): ResponseInterface
             {
                 /** @var string $body */
-                $body = file_get_contents(dirname(__DIR__, 2).'/test_data/public_suffix_list.dat');
+                $body = file_get_contents(dirname(__DIR__, 2).'/test_data/tlds-alpha-by-domain.txt');
+
                 return new Response(200, [], $body);
             }
         };
@@ -52,10 +52,10 @@ final class PublicSuffixListRemoteStorageTest extends TestCase
             }
         };
 
-        $storage = new PublicSuffixListRemoteStorage($client, $requestFactory);
-        $psl = $storage->getByUri('http://www.example.com');
+        $storage = new RootZoneDatabaseRemoteStorage($client, $requestFactory);
+        $rzd = $storage->getByUri('http://www.example.com');
 
-        self::assertInstanceOf(Rules::class, $psl);
+        self::assertInstanceOf(TopLevelDomains::class, $rzd);
     }
 
     public function testItWillThrowIfTheClientCanNotConnectToTheRemoteURI(): void
@@ -74,10 +74,10 @@ final class PublicSuffixListRemoteStorageTest extends TestCase
             }
         };
 
-        self::expectException(UnableToLoadPublicSuffixList::class);
-        self::expectExceptionMessage('Could not access the Public Suffix List URI: `http://www.example.com`.');
+        self::expectException(UnableToLoadRootZoneDatabase::class);
+        self::expectExceptionMessage('Could not access the Root Zone Database URI: `http://www.example.com`.');
 
-        $storage = new PublicSuffixListRemoteStorage($client, $requestFactory);
+        $storage = new RootZoneDatabaseRemoteStorage($client, $requestFactory);
         $storage->getByUri('http://www.example.com');
     }
 
@@ -97,11 +97,11 @@ final class PublicSuffixListRemoteStorageTest extends TestCase
             }
         };
 
-        self::expectException(UnableToLoadPublicSuffixList::class);
-        self::expectExceptionMessage('Invalid response from Public Suffix List URI: `http://www.example.com`.');
+        self::expectException(UnableToLoadRootZoneDatabase::class);
+        self::expectExceptionMessage('Invalid response from Root Zone Database URI: `http://www.example.com`.');
         self::expectExceptionCode(404);
 
-        $storage = new PublicSuffixListRemoteStorage($client, $requestFactory);
+        $storage = new RootZoneDatabaseRemoteStorage($client, $requestFactory);
         $storage->getByUri('http://www.example.com');
     }
 }
