+++ src/TopLevelDomainsTest.php
@@ -13,17 +13,10 @@
 
 declare(strict_types=1);
 
-namespace Pdp\Tests;
+namespace Pdp;
 
 use DateTimeImmutable;
 use DateTimeZone;
-use Pdp\Domain;
-use Pdp\InvalidHost;
-use Pdp\PublicSuffix;
-use Pdp\RootZoneDatabaseConverter;
-use Pdp\TopLevelDomains;
-use Pdp\UnableToLoadRootZoneDatabase;
-use Pdp\UnableToResolveDomain;
 use PHPUnit\Framework\TestCase;
 use TypeError;
 use function dirname;
